# swagger_server/models/error_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.error_type_enum import ErrorTypeEnum  # noqa: F401,E501
from swagger_server import util


class ErrorResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, code: str=None, type: ErrorTypeEnum=None, message: str=None, details: str=None):  # noqa: E501
        """ErrorResponse - a model defined in Swagger

        :param code: The code of this ErrorResponse.  # noqa: E501
        :type code: str
        :param type: The type of this ErrorResponse.  # noqa: E501
        :type type: ErrorTypeEnum
        :param message: The message of this ErrorResponse.  # noqa: E501
        :type message: str
        :param details: The details of this ErrorResponse.  # noqa: E501
        :type details: str
        """
        self.swagger_types = {
            'code': str,
            'type': ErrorTypeEnum,
            'message': str,
            'details': str
        }

        self.attribute_map = {
            'code': 'code',
            'type': 'type',
            'message': 'message',
            'details': 'details'
        }
        self._code = code
        self._type = type
        self._message = message
        self._details = details

    @classmethod
    def from_dict(cls, dikt) -> 'ErrorResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ErrorResponse of this ErrorResponse.  # noqa: E501
        :rtype: ErrorResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def code(self) -> str:
        """Gets the code of this ErrorResponse.

        Application error code  # noqa: E501

        :return: The code of this ErrorResponse.
        :rtype: str
        """
        return self._code

    @code.setter
    def code(self, code: str):
        """Sets the code of this ErrorResponse.

        Application error code  # noqa: E501

        :param code: The code of this ErrorResponse.
        :type code: str
        """

        self._code = code

    @property
    def type(self) -> ErrorTypeEnum:
        """Gets the type of this ErrorResponse.


        :return: The type of this ErrorResponse.
        :rtype: ErrorTypeEnum
        """
        return self._type

    @type.setter
    def type(self, type: ErrorTypeEnum):
        """Sets the type of this ErrorResponse.


        :param type: The type of this ErrorResponse.
        :type type: ErrorTypeEnum
        """

        self._type = type

    @property
    def message(self) -> str:
        """Gets the message of this ErrorResponse.

        Short error explanation  # noqa: E501

        :return: The message of this ErrorResponse.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this ErrorResponse.

        Short error explanation  # noqa: E501

        :param message: The message of this ErrorResponse.
        :type message: str
        """

        self._message = message

    @property
    def details(self) -> str:
        """Gets the details of this ErrorResponse.

        Short error detail  # noqa: E501

        :return: The details of this ErrorResponse.
        :rtype: str
        """
        return self._details

    @details.setter
    def details(self, details: str):
        """Sets the details of this ErrorResponse.

        Short error detail  # noqa: E501

        :param details: The details of this ErrorResponse.
        :type details: str
        """

        self._details = details