# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: IPL Tracking - Microservice API
  description: API description for Tracking microservice.
  version: 1.0.0
servers:
- url: http://localhost:8080/tracking
  description: Localhost - BT3 Dev
...
paths:
  /customers:
    get:
      tags:
      - Customer
      summary: Get Customers list
      description: This operation is used to retrieve a list of Customers.
      operationId: list_customers
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCustomersResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.customer_controller
    post:
      tags:
      - Customer
      summary: Create new Customer
      description: This operation is usedto create a new Customer on System.
      operationId: create_customer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'
        "400":
          description: Parameters missing in request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: There is already a customer registered with same name or mail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.customer_controller
  /customers/{customerId}:
    get:
      tags:
      - Customer
      summary: Get a single Customer's info
      description: This operation is used to retrieve the details of a specific device.
      operationId: get_customer
      parameters:
      - name: customerId
        in: path
        description: Unique identifier of the Customer in the database
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/UUID'
        example: 0af401df-7d7a-1f36-817d-7b0a058d0003
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCustomerResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.customer_controller
    delete:
      tags:
      - Customer
      summary: Delete Customer
      description: This operation is delete a Customer.
      operationId: delete_customer
      parameters:
      - name: customerId
        in: path
        description: Unique identifier of the Customer in the database
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/UUID'
        example: 0af401df-7d7a-1f36-817d-7b0a058d0003
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Delete a customer that has vehicles is not allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.customer_controller
    put:
      tags:
      - Customer
      summary: Update Customer's attributes
      description: This operation is used to update some of the Customer's attributes.
      operationId: update_customer
      parameters:
      - name: customerId
        in: path
        description: Unique identifier of the Customer in the database
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/UUID'
        example: 0af401df-7d7a-1f36-817d-7b0a058d0003
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
        required: true
      responses:
        "204":
          description: No Content
        "400":
          description: "Parameters missing in request or request is malformed. For\
            \ example, the mail is not in the correct format."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: There is already a Customer registered with same name or email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.customer_controller
...
components:
  schemas:
...
    GetCustomerResponse:
      required:
      - customerId
      - mail
      - name
      - phone
      type: object
      properties:
        customerId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Complete Customer name
          example: Foo bar
        phone:
          type: string
          description: Telephone number
          example: "351912000111"
        mail:
          type: string
          description: E-mail address
          format: email
          example: foo@bar.com
      example:
        mail: foo@bar.com
        phone: "351912000111"
        customerId: 0af401df-7d7a-1f36-817d-7b0a058d0003
        name: Foo bar
    CreateCustomerRequest:
      required:
      - mail
      - name
      - phone
      type: object
      properties:
        name:
          type: string
          description: Complete Customer name
          example: Foo bar
        phone:
          type: string
          description: Telephone number
          example: "351912000111"
        mail:
          type: string
          description: E-mail address
          format: email
          example: foo@bar.com
    CreateCustomerResponse:
      required:
      - customerId
      type: object
      properties:
        customerId:
          $ref: '#/components/schemas/UUID'
      example:
        customerId: 0af401df-7d7a-1f36-817d-7b0a058d0003
    ListCustomersResponse:
      required:
      - content
      - totalResults
      type: object
      properties:
        content:
          type: array
          description: list of paged items
          items:
            $ref: '#/components/schemas/GetCustomerResponse'
        totalResults:
          type: integer
          description: total number of records
          format: int64
          example: 50
      description: generic paged response
      example:
        totalResults: 50
        content:
        - mail: foo@bar.com
          phone: "351912000111"
          customerId: 0af401df-7d7a-1f36-817d-7b0a058d0003
          name: Foo bar
        - mail: foo@bar.com
          phone: "351912000111"
          customerId: 0af401df-7d7a-1f36-817d-7b0a058d0003
          name: Foo bar
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
          description: Complete Customer name
          example: Foo bar
        phone:
          type: string
          description: Telephone number
          example: "351912000111"
        mail:
          type: string
          description: E-mail address
          format: email
          example: foo@bar.com
...
    ErrorResponse:
      required:
      - errorCode
      - errorMessage
      - errorType
      type: object
      properties:
        code:
          type: string
          description: Application error code
          example: XYZ0001
        type:
          $ref: '#/components/schemas/ErrorTypeEnum'
        message:
          type: string
          description: Short error explanation
          example: Unauthorized
        details:
          type: string
          description: Short error detail
          example: Parameter X has invalid value
    ErrorTypeEnum:
      type: string
      description: Error Type
      enum:
      - PERSISTENCE
      - BUSINESS
      - COMMUNICATION
      - UNKNOWN
    UUID:
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      type: string
      example: 0af401df-7d7a-1f36-817d-7b0a058d0003
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_basicAuth